+++ src/com/twitter/intellij/pants/metrics/PantsMetrics.java
@@ -1,7 +1,7 @@
 // Copyright 2016 Pants project contributors (see CONTRIBUTORS.md).
 // Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-package com.twitter.intellij.pants.components.impl;
+package com.twitter.intellij.pants.metrics;
 
 import com.google.common.base.Stopwatch;
 import com.intellij.openapi.application.ApplicationManager;
